added base sensor and dummy sensor
